Creating executor interface
