+++ executors/executor.py
@@ -1,0 +1,6 @@
+class Executor:
+    def setOn(self):
+        pass
+
+    def setOff(self):
+        pass
